+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
 - Clean up the docs.
+- Allow numeric indexing on OME-TIFF loader.
 
 ## 0.2.2