+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 ### Changed
 
+## 0.2.3
+
+### Added
+
+### Changed
+
 - Add OME-TIFF to Covid-19 data description.
 - Clean up the docs.
 - Allow numeric indexing on OME-TIFF loader.